Use go/loader instead of custom logic
Get rid of a lot of complexity by consistently using go/loader instead
of a combination of manual source parsing and reading of gc export data.

Benefits:

- easier code
- cgo support
- no staleness issues anymore

Downsides:

- noticably slower because the stdlib needs to be parsed

Behaviour changes, to address downsides:

- when linting more than one package, all packages will now be treated
as a single program. This avoids having to reparse the stdlib several
times, but can lead to failures if any of the packages is broken.