ssa, staticcheck: emit pseudo instruction for assignment to the blank identifier
Assigning to the blank identifier has no effect. When a user assignes to
it, however, it is often an indication of knowingly and purposefully
discarding a value, which is why we shouldn't flag variables assigned to
it as unused, unnecessary or the like.

By introducing a new pseudo-instruction BlankStore we ensure that values
used only in the context of the blank identifier have a non-zero number
of referrers.

Closes dominikh/go-staticcheck#73