all: refactor lint framework to work on "jobs"
The old model, which we inherited from golint, operated on a file by
file basis. For most tools, this would require workarounds to operate on
packages or programs.

The new model has the notion of a "job", which describes the entire
program, with all its files, packages and functions. The individual
tools and checks can choose to iterate over whichever of the sets they
want to.

This required shuffling a bunch of APIs around and getting per-file
information indirectly, via cached node to file and file to package
mappings.