lint: support linter directives for ignoring problems
This commit adds a "//lint:ignore" linter directive for ignoring
problems on specific lines.

In the past, we've refused adding such a mechanism, due to the Go team's
distaste for tool directives in source code. However, there are no good
alternatives, and false positives unfortunately happen.

In the past, we encouraged file-wide ignore patterns that are managed
out of band. This, however, leads to broad patterns that aren't
maintained well.

We believe that in-band directives don't have to be too intrusive if
some rules are followed. We will consider them a last resort, when
neither changing the code nor the linter is a viable option.
Furthermore, the directive requires adding a human-readable reason for
adding the directive. This way, the comment will provide actual value to
the human reader.

Additional mechanisms that will be added in the future will help make
sure that outdated directives get removed, instead of cluttering the code.

The directive introduced in this commit takes the form of

  //lint:ignore CheckID1[,CheckID2,...,CheckIDN] These checks triggerfalsely because ...

A directive is associated with the closest node; either the node that
ends on the same line as the comment begins, or the node that follows.
Essentially, it follows the rules of go/ast CommentMaps.

This style has the downside of showing up in godoc when associated with
functions or global variables. However, the alternative, having
directives apply to the line they appear on, has its own downsides: only
one directive can appear per line and reasons can't word-wrapped.

Closes gh-205