lint: report ineffective linter directives
A common problem with linter directives is that people don't maintain
them. As code changes, linter directives become unnecessary but remain
in the code.

We can remedy this situation by making unnecessary directives a problem
that we report. Each tool knows all of its checks and can report
directives that were never triggered.

Some things have to be kept in mind, however:

- This should only be done with line-based directives. File-based
directives can be cautionary (e.g. "this auto-generated code often
contains dead code, ignore it when it happens"). Glob-based patterns may
or may not hit depending on which packages are being checked.

- We should not flag directives that might match a disabled check.
Special care must be taken for directives that use globs.

- We should not flag directives of other tools. Since all IDs are
prefixed with a namespace, and each tool knows its own namespace, this
shouldn't be a problem.

- We mustn't short-circuit in the ignore check, as multiple
ignores (glob ones, file ones, and line ones) may overlap and match the
same problem.

Closes gh-210