all: clean up handling of //line directives
The //line compiler directive can be used to change the file
name and line numbers associated with code. This can, for
example, be used by code generation tools. The most prominent
example is 'go tool cgo', which uses //line directives to refer
back to the original source code.

In the context of our linters, we need to treat these
directives differently depending on context. For cgo files, we
want to honour the directives, so that line numbers are
adjusted correctly. For all other files, we want to ignore the
directives, so that problems are reported at their actual
position and not, for example, a yacc grammar file. This also
affects the ignore mechanism, since it operates on the position
information stored within problems. With this implementation, a
user will ignore foo.go, not foo.y