unused: support the NoCopy sentinel value
The Go standard library internally introduced the idea of a "no copy"
sentinel value that structs could use. The presence of the sentinel
would signal to go vet that copying of the struct is forbidden.

This idea leaked to other people. For example, CockroachDB has its own
util.NoCopy type that is used in various types. This triggered spurious
unused field warnings.

We adopt go vet's notion of the sentinel by handling the following kind
of type specially: A struct type with no fields and exactly one method
func Lock(). Fields of this type will always be considered used.

Note that ideally we'd check that the Lock method is empty, but we're
not currently doing that.

Closes gh-214