lint: use unadjusted line number if we're unsure
In the past, we defaulted to using the adjusted line number if we
couldn't find a build.Package for a given file – a case that could only
be hit on edge cases.

With the new loader, however, packages created from files (e.g.
'staticcheck foo.go') will not have associated build.Packages. Since
this is the only realistic case for which no build.Package is available,
and since this mode of operation doesn't support cgo anyway, we should
default to using the unadjusted line number.