simple: remove S1013
S1013 has been the cause for some contention. Respected developers
disagree with its suggestions, primarily on account of breaking
consistency with surrounding code.

While we could try to detect some forms of consistency in surrounding
code, we'd never be able to catch all of them. Since the check isn't
always right, and didn't provide drastic improvements in the first
place, it's best to remove it.

Closes gh-273
Closes gh-266