staticcheck: optimize finding deprecated objects
Instead of finding all members and trying to map them back to the AST,
walk the AST directly. By carefully only traversing those subtrees that
are required for finding deprecated objects, as opposed to the entire
tree, the overall CPU and memory usage can be reduced.

For 'staticcheck std', the overall runtime is reduced by 12% and the max
memory usage is reduced by 9%.