Revert "ssa: explicitly initialize blank variables with implicit zero value"
We made this change so that code like `var _ B` would generate code for
unused to pick up on and consider B used. However, the change made
doesn't work for `var _ *T = nil`, because that generates a blank store
of an untyped nil.

We could change this to store a typed nil, but at that point we'd
effectively be lying about the code. Instead, we'll add AST-based code
to unused to handle blank variable declarations.

This reverts commit 3e913d42fb6282d2f42e40eda15c3a2ee564d3f8.