unused2: track used identifiers across files
A single file can be part of multiple packages, for example non-test and
test variants. If an identifier is used in one of the packages but not
the others, we shouldn't report it as unused. This most often occurs
with functions defined in a non-test file but only used in tests.

In the future we will want to add a flag to disable this, so that code
only used by tests can optionally be flagged.