Delete errcheck
Writing a useful version of errcheck requires whole-program analysis and
points-to analysis. It would be incredibly slow and memory intensive for
real world applications. We can't possibly run it as part of
staticcheck, and the Go community is not open to static analysis that
needs hours or days to complete.