lint: remove map from import paths to packages
This mapping is not actually 1 to 1. Test variants and vendoring mean
that the same path can refer to many different packages. Especially test
variants in the standard library make this very complex because of
transitive test dependencies.

Especially our only use of the map, an Implements function, is
inherently unreliable. We can't go from "fmt.Formatter" to the correct
package and type. We don't know _which_ fmt package a value may end up
in, which means we don't know which fmt.Formatter to apply.

These checks will have to be implemented differently. For example,
fmt.Stringer can be implemented soundly by looking for a correctly
formed String method. fmt.Formatter, on the other hand, will have to
rely on less precise string matching and risk matching too many types.