ssa: replace Langauer & Tarjan algorithm with optimized iterative one
We implement a well-optimized, iterative algorithm for computing
dominance, based on the algorithm shown in A Simple, Fast Dominance
Algorithm, by Cooper et al. While the paper's claims of improved
performance over even well-optimized LT have been disproven (see
Finding Dominators in Practice, by Georgiadis, Tarjan, and Werneck),
the algorithm is still competitive and achieves similar results on the
kinds of graphs we encounter as CFGs for Go programs. While variants
of LT are superior for very large graphs, our graphs tend to be orders
of magnitude smaller.

In light of that, we opt for using a much simpler algorithm. It uses
less code and is easier to understand. It should also have a lower
memory footprint, though neither algorithm dominates memory usage of
SSA generation.

We apologize profusely to the original author of the go/ssa package,
who must've spent a considerable of time on the implementation we're
throwing away.