unused: take special care when deduplicating types
Deduplicating types requires some special considerations.

The first issue is that multiple functions may share the same
signature, but we're still adding individual receivers and parameters
to the graph. We must be careful to not miss some of them when
ignoring nodes.

The second issue involves type identity of interface types.
types.Identical considers the underlying types of I1 and I2 identical
in the following code:

    type I1 interface {
      Fn()
    }

    type I2 interface {
      I1
    }

Thus, if we first process I1's underlying type and then I2's, we would
miss the fact that I2 uses I1. We work around this by handling
interfaces specially in our own implementation of Implements.
Specifically, we fall back to pointer equality. This will cause the
graph to grow more, but it should have no effect on correctness.

This commit also changes graph edge labels from strings to enums. This
change was part of an earlier experiment. It isn't required for the
correctness of this change, but we liked it enough to keep it, anyway.