Fix atomic accesses on 32-bit platforms
When atomically accessing 64-bit values on 32-bit platforms, they must
be aligned on 64-bit boundaries, and we're responsible for ensuring
this alignment.

For lint.Stats, we switch to 32-bit fields, so that we don't have to
be careful about how we allocate the struct. 32-bit should be plenty
to keep track of packages and problems.

In the trivial cases, we've moved the 64-bit fields to the top of the
struct.