staticcheck: don't flag uint <= 0
While it is true that uint <= 0 can be written more specifically as
uint == 0, this is not a helpful suggestion in reality. A lot of people use
uint <= 0 as a defensive measure, in case uint ever becomes signed. Also,
unlike all the other warnings made in the check, uint <= 0 is neither a
tautology nor a contradiction, it is merely less precise than it could
be. As such, changing flagged code won't fix any bugs. A lot of people
considered this warning to be noise.

Closes gh-534