staticcheck: don't flag "unread" variable used in switch with no branches
Avoid flagging code that reads as such:

    key := in.UnsafeString()
    switch key {
    default:
    	in.SkipRecursive()
    }

While the value of `key` is indeed useless, this pattern primarily
shows up in generated code, making it a noisy positive.

For the moment we choose to ignore these unread values in all code,
not just generated code, based on the rationale that it looks more
used than it looks unused.

Closes gh-533