staticcheck: allow use of time.Tick in Cobra commands
Note that our check for cobra is trivial, it only checks if cobra is
being imported. We do not check if commands are being defined. The
assumption is that there are no reasons for importing cobra other than
to implement commands. Even if that is not the case, it seems unlikely
that a package imports cobra, uses time.Tick and is not a command.

Closes gh-535