internal/gosmith: add a copy of gosmith
This is almost a verbatim copy of
upstream (https://github.com/dvyukov/gosmith) with the only
modification being that global state was converted into a struct.

The plan is to further modify gosmith to produce instances of
go/packages.Package instead of writing to the file system. We will
then use it in fuzz tests of staticcheck.