staticcheck: in SA4014, do much less work in the common case
The majority of if statements do not have an else-if branch, and we
don't have to render the sole condition to figure out that there are
no duplicates. We don't even have to look at the if statement if it
has no else branch.

Furthermore, there is no point in collecting all conditions, only to
determine if one of them has side effects – we can bail early the
moment we see a condition with a side effect. The same goes for init
statements.

This speeds up the check by a factor of 7.

Updates gh-601
Updates gh-606