code: make MayHaveSideEffects actually be exhaustive
In f3ea247d8f995d055e7323e61cbf7be7349d41b3 we added an "exhaustive"
check for side effects in expressions. Unfortunately, that
implementation wasn't exhaustive after all, missing a handful of node
types. None of these are commonly seen in conditionals, which is why
the panic was unlikely to occur.