ssa: detect control flow due to syscall.Exit, runtime.Goexit and panics
Future SSI based analyses will depend more heavily on control flow.
For example, a future nil pointer check will flag unsafe pointer
dereferences that succeed a non-terminating nil pointer check. To
avoid false positives, we need to be aware of the control flow
implications of terminating the process or goroutine.

We implement an understanding of syscall.Exit, runtime.Goexit and
panic in called functions, and a way of propagating the "will exit"
and "will unwind" information between functions and packages.

We introduce a new instruction Unreachable that acts as an
unconditional jump to the exit block, but implies that the exit block
will not actually execute. In other words, it signals that the
previous function call will not return, due to a call to syscall.Exit.
For runtime.Goexit as well as calls to panic, we emit an ordinary jump
to the exit block, as these calls still cause defers to run.

Having to propagate information between packages means that SSA
building is no longer trivially parallelizable. However, our primary
use of go/ssa is via go/analysis, which already implements its own
parallelism based on the dependency graph, and so we simply remove
support for parallel building from go/ssa.