Rename ssa package to ir, update all references
In olden days staticcheck used upstream's go/ssa package. One day we
decided that we needed to make small adjustments to the SSA IR and we
forked the package. We had restricted ourselves to small modifications,
staying close to the upstream representation. All of that changed when
we decided to move past SSA, by implementing SSI. Yet the package was
still called ssa, no longer a reflection of its true nature. This
commit fixes that.

This may seem like a very invasive and unnecessary commit. It touches
the majority of files in the repository, making what seems like a
pedantic change. I do believe, however, that this change is beneficial,
making it clear that our IR is neither identical to upstream's
go/ssa – a mistaken assumption a certain metalinter has made in the
past – nor restricted to SSA in general. As development on
staticcheck progresses, we may experiment with other modifications to
the IR. What was SSA yesterday is SSI today and may become something
else entirely tomorrow. Hence the generic name of 'ir'.

Henceforth let it be known that our IR is not limited to SSA, nor SSI
nor SSU. It is our IR. It is the best IR.