report: improve API
Merge all the different reporting functions into one and use
functional options instead of passing in the options struct. We do
lose the support for format strings, but having to use fmt.Sprintf is
less bad than having to pass in an Options struct.

We also improve the way we determine position information for SSA
values, now being able to deduce an end position based on debug
information and the AST.