ir: overhaul tracking of position information
For our analyses, we need start and end position information for
diagnostics, for the best possible user experience in editors.

The original go/ssa package, however, only tracks a single position
per SSA node, with no real way of getting an instruction's End
position. DebugRefs allow mapping between expressions and values, but
this doesn't apply to non-value instructions. The other alternative,
using PathEnclosingInterval to look up an AST node given a position
information is too costly.

We could have extended our IR to also track End positions, but this
would lead to another problem: different analyses want different
ranges for the same AST nodes. For example, when flagging an
unnecessary 'if' around a piece of code, we want to only flag the 'if'
keyword and the condition, not the entire body. In other cases,
however, we may want to flag the body, too.

Thus, instead of tracking position information at all, we track which
AST node was responsible for the creation of an instruction. We call
this the "source".

The original go/ssa package has been designed in a way that avoids
pinning the AST in memory. This is one of the reasons why DebugRefs
are optional, and why SSA instructions hold no references to AST
nodes. However. our IR is only used in contexts in which the AST is
needed, anyway, so we do not incur any extra cost by referencing it in
our instructions.

go/ssa also made sure that each position was used for at most one
instruction, and that it was possible to uniquely look up an
instruction from an AST node's position. Our own analyses never
depended on this attribute. On the other hand, the uniqueness of
position information meant that some instructions would have no
position information at all. For example, a selector expression 'a.x'
may involve multiple implicit selectors 'a.b.c.x' – and because these
aren't present in the source, no unique positions exist to associate
with the field lookups emitted in the IR. Our IR, however, has no
issue with making a single AST node be the source of many
instructions. In the same selector example, all implicit selectors
would have the same source AST node, the explicit selector.