Remove sponsors from README
Originally we had no website and only one sponsor, so we put their
logo in the README. Eventually we got a website, and promised sponsors
that we'd put their logos on the website. However, we also added them
to the README, for fairness. Over time we have gotten quite a few
sponsors, and they change somewhat frequently, which means that

- too many files in this repository are just images
- too many commits are just about changing the list of sponsors
- the README is getting quite long, and due to limited styling, looks
  really cluttered.

To tidy things up, we'll only display logos on the
website (staticcheck.io) from now on. That's in line with what we
promise people when they become sponsors. We still prominently mention
the existence of sponsors, at the top of the README, and link to a
list of all of them.

I hope that this doesn't upset anyone.