staticcheck: the %x and %X verbs now support floats and complex numbers
Go 1.13 changed fmt to allow the use of floats and complex numbers
with the %x and %X verbs. Adjust our check accordingly.

Note that we do not check for the Go version, and assume that %x
always behaves as in 1.13. This only causes false negatives, and it's
probably not worth complicating the code to fix them.