ir: don't allocate space for builder state for body-less functions
We primarily use ir in the context of go/analysis and the buildir
pass, which creates a new IR program for each package under analysis.
For large code bases, this creates millions of instances of Function,
created from export data. These functions do not need any of the
transient builder state to be built.

Moving this state into a separate struct that gets allocated
conditionally saves approximately 50% of space for these repeated
allocations, or a reduction of 10 GB when analysing cockroachdb.