facts: handle overlapping prefix/suffix
We have a generated file which starts with the comment:

```
// Code generated DO NOT EDIT.
```

This causes staticcheck to panic, because we have an overlapping prefix and
suffix when checking for these magic strings, i.e.:

```
"// Code generated "
                 " DO NOT EDIT."
```

Trying to take a slice where end is before start subsequently panics.

This commit makes the "parse" non-overlapping, at the cost of replacing the
text switch lookup with `strings.HasPrefix`. A plausible alternative would
be to do an explicit bounds test instead - I haven't bothered thinking too
hard / benching these, though.