facts: handle overlapping prefix/suffix
We have a generated file which starts with the comment:

```
// Code generated DO NOT EDIT.
```

This causes staticcheck to panic, because we have an overlapping prefix and
suffix when checking for these magic strings, i.e.:

```
"// Code generated "
                 " DO NOT EDIT."
```

Trying to take a slice where end is before start subsequently panics.

Add an explicit check to avoid the out of bounds access.

Closes: gh-716 [via git-merge-pr]