Skip excessively large packages
A common (albeit terrible) way of bundling assets with Go packages is
to convert them into Go source files, usually some variation of a
giant byte slice or string containing binary data. These files can be
as large as hundreds of megabytes for some people, and parsing,
type-checking and building IR form for them is an extreme burden with
little reward. If we assume that assets are confined to their own
packages, with no other logic mixed in, then we really don't care
about analyzing these packages. No problems will be found, and no
useful facts will be contributed to the analysis of dependents.

We set the limit at 50 MB for now. We initially tried with 1 MB and 10
MB, which quickly ran into generated code for protobufs, and machine
converted code respectively. We may need to tweak the limit further in
the future.