lintcmd: print reference to -explain check to stderr
Separate diagnostics from help output. The output of the plain
formatter is supposed to be machine readable, so don't emit text meant
strictly for humans.

For the stylish formatter we could've gone either way, because its
output is not meant to be processed by machines, but we copy the
behavior of the plain formatter for the sake of consistency.