Cache staticcheck export data between CI runs
This doesn't really make our CI run any faster, because running the
tests takes much longer than running staticcheck, even uncached.
However, our approach to caching should serve as a good reference to
other people using staticcheck, hopefully with faster tests.

We experimented with caching test results, but that was complicated by
how cache keys for tests are computed. The test binary generates a log
of interactions with the environment, such as statting files, and the
environment gets captured in the cache key. In the case of stat, this
includes the ctime and mtime, which will be different for every run of
CI. We could set all ctimes and mtimes to a fixed timestamp, but that
would require different scripts for different operating systems – or a
helper utility.

We also experimented with caching $GOPATH/pkg/mod, but action/setup-go
doesn't make it easy to figure out what GOPATH is, and downloading our
dependencies is not enough of a bottleneck to warrant complicating our
workflows further.