go/ir: improve analysis of functions that don't return
This change adds support for functions that abort control flow, but
use a mixture of exiting the process and panicking. In the past, we
would only detect functions that either always exit or always panic.

That is, we now correctly detect that the following function
affects control flow:

	func fn(b bool) {
		if b {
			panic("")
		} else {
			syscall.Exit(1)
		}
	}

This change requires the introduction of a new intrinsic,
ir:noreturnWasPanic, which we use to encode the runtime behavior of
functions that may either panic or exit.

As part of these changes we happen to fix a bug: given two functions A
and B, where A panics unconditionally and B calls A, but recovers from
the panic, we would previously incorrectly mark B as always panicking
also, ignoring the fact that it recovered from the panic.

This change was triggered by os.Exit in Go 1.16 conditionally
panicking or exiting. Instead of special-casing os.Exit, I decided to
improve handling of such functions, so that user-written code may
benefit from the improved analysis.

Closes gh-872