staticcheck: don't use AST walking to find node corresponding to IR call instruction
Before c14c261fd1fd7dec2e23b44bbf8c7b49cbd4e239, the only way to map
from an IR instruction to an AST node was by using position
information and walking the AST, usually by using
astutil.PathEnclosingInterval.

Back then, the position information stored with instructions was
designed to map uniquely to AST nodes. For example, in
'fn1().fn2().fn3()`, the call of fn3() would have its position on the
dot immediately before fn3. Since the change, however, The start
position will be at the beginning of the sequence of expressions.
Using that position to find the correct call is wrong and will yield
the wrong node. For defer and go statements, we might even
accidentally look at the surrounding code.

In the best case, this would lead to incorrect positions in our
output. In the worst case, it would lead to out of bounds crashes due
to a mismatch in argument counts.

Note that many IR instructions may map to the same AST node, for
example because of method calls of embedded methods. This, however,
shouldn't be a problem here, as the AST node should always correspond
to the one we're interested in.

Closes gh-873