staticcheck/doc.go: SA4008 flags loops with unconditional exits
The warning SA4008 "variable in loop condition never changes" is
triggered when a loop has an unconditional exit, such as a break
or panic. Since the warning is logged on the line of the loop
which could be far from the unconditional exit, it some cases at
first glance it appears this warning is wrong. To help, add
documentation showing that this triggers for unconditional exits.
This should help the next person who runs into this.

This was the cause of a "false positive" bug that was closed as
working as intended: https://github.com/dominikh/go-tools/issues/868