lintcmd: don't mention the -explain flag
Mentioning the flag on every run that produces diagnostics has turned
out to be too annoying. Most users will remember that the flag exists
after the first time we've told them about it. After that, we're just
making output longer and more noisy for people. This is especially
evident when run from inside editors.

Instead of adding more heuristics or flags to control this behavior,
remove it. We'll have to figure out a better way of teaching our users.

This reverts 9cc924ae9826f784f093cce37c36a4dd01b48371,
9b21780a6dea54d84dd05263ec284266081c9679 and 08a27f44447472c7758985be6d8e34fea93fbeec.