staticcheck: improve extractConsts helper
In commit ae6041020cd7fb7ad9a5d6c251130b93487d0a73 we changed
extractConsts to not collect all distinct constants from phi nodes. We
didn't notice that the function no longer returned more than one
constant and didn't need to return a slice anymore.

In this change, we change the signature and name, and replace all
looping with a single if != nil check.

Additionally, we introduce the extractConstExpectKind helper, which
takes care of checking that the constant is of the right kind, and not
untyped nil. This helper removes a lot of boilerplate.

(cherry picked from commit 173553237fc1eb9293e0810c58bba23c08407546)