all: combine x/tools's typeutil and our typeutil
Originally, we vendored a copy of x/tool's typeutil, with a modified
Map implementation. That Map implementation was deleted in
41bedbdbf9e6154d6edae49010d46bd375b076f2, rendering the remainder of
the vendored copy unnecessary. We also weren't very consistent in
using one or the other package.

Give typeutil the same treatment as astutil: re-export the specific
bits of x/tools that we need, and only use our own package everywhere.