Begin work on a SARIF formatter
This adds incomplete support for a SARIF formatter. Many bugs still
need fixing.

To provide the formatter the required information about analysers, we
introduce a new lint.Analyzer type which provides access to our
structured documentation, as opposed to the single string that
is go/analysis.Analyzer.Doc.

We introduce the use of minimal Markdown (primarily backticks for
inline code and indentation for blocks of code) to render nicer
documentation. The lint.Markdownify helper turns \' into ` – a much
nicer way of writing backticks in raw strings. We limit yourselves to
a subset of Markdown that makes sense when read as plain text, so that
we can use it directly in terminal output.