unused: revert to using a typeutil.Map
In 41bedbdbf9e6154d6edae49010d46bd375b076f2 we incorrectly decided
that unused no longer needed to deduplicate types. For the most part,
this was correct. However, it failed to account for pointer types.
go/types does not guarantee that types.NewPointer(T) ==
types.NewPointer(T). This caused parts of the graph to be incorrectly
unreachable.

This commit reverts the aforementioned commit, but puts our custom
Map in its own package, only to be used by the 'unused' check.

Closes gh-942