analysis/report: better handling of sourcers with no node
Some IR values do not have a valid Source, most prominently constants.
Don't panic when we can avoid it: when trying to emit related
information, it makes more sense to drop the related information on
the floor than to panic. We still panic for the toplevel diagnostic,
because it makes no sense to print a diagnostic for an unknown
location. It is an explicit panic now, however, so we get a more
helpful message.

Closes gh-972