cmd/rdeps: delete rdeps
The move to Go modules has greatly diminished the usefulness of the
rdeps tool, as there is no longer a universe of packages (GOPATH) to
inspect, and because we've never updated the tool to work with modules
in the first place.

On top of that, a much more powerful tool – goda – has emerged, which
allows for far more advanced queries on dependency graphs in the scope
of a module.

Closes gh-495