stylecheck: don't complain about empty documentation
Our checks that validate documentation on exported objects only flag
malformed documentation, not missing documentation.

As a special case in (*ast.CommentGroup).Text, tool directives such as
//go:noinline get stripped from the documentation. This can lead to
documentation that is non-nil but empty, if the comments only
consisted of such directives.

However, we don't want to flag

    //tool:directive
	type X int

for having a malformed comment.

With this change, we will treat empty documentation the same as
missing documentation and ignore it. This does mean that we won't flag
the following:

    //
	type X int

That's really a side-effect of how we skip over directives in
comments, and arguably we might want to treat "just directives"
differently from "empty comment", but it is probably not worth the
added complexity.