quickfix: delete QF1000
The upstream position is that one should always use strings.Index when
dealing with literals, and we don't want to offer a suggestion that
goes against that. IndexByte has a small performance improvement over
Index, but this should be addressed by improvements to strings.Index
and the compiler.

We could still recommend to use IndexByte when not using literals, but
that would be entirely new code, not part of the existing check.

It is not clear what to do for the bytes package, but it is probably
best to err on the side of not making any suggestions.

Closes gh-1007