SA4010: don't flag appends to slices that might have aliased backing arrays
Multiple slices may refer to the same array, which means the effect of
append can be observed even if append's result gets discarded.

We can only be sure about slices that refer to memory allocated in the
same function and that don't get used in any way that might retain
them, other than the specific chain of appends we're interested in.

Closes gh-545