unused: don't ignore fields and methods of aliased type in imported package
When ignoring an alias, we don't want to ignore the fields and methods
of the aliased type if that type is in a different package. Not only
is it unnecessary, it also breaks the invariant that we only use
objects we've already seen.

Closes gh-1073

(cherry picked from commit 6c9d8ed69b182b232989e3b972da00556130f856)