website: add staticcheck.io website
This adds a new website for Staticcheck. It uses Hugo[1] and the
Docsy[2] theme and is fully open source, unlike the old website.

The only files omitted from this commit are the list of sponsors and
sponsor logos. These are kept out of tree, to keep the repository
small and to avoid unnecessary commits as sponsors come and go.
The website still builds fine without those files, it will only be
missing the sponsors bar at the bottom, as well as the list of
sponsors on the Sponsors page.

To make generating the check documentation easier, lint.Documentation
has gained two new fields, "Before" and "After". These contain before
and after code samples, which were previously part of the "Text"
field.

The website is its own Go module, to prevent its inclusion in the
honnef.co/go/tools module's ZIP. It's part of this repository,
however, so that changes to Staticcheck and its documentation stay in
sync.

[1]: https://gohugo.io/
[2]: https://www.docsy.dev/