Implement a more efficient test runner
We've previously used x/tools/go/analysis/analysistest for our tests.
However, that package is not suited to checking hundreds of analyzers,
as it runs a single analyzer at a time. As a consequence, it has to
load dependencies again and again for each analyzer, which gets costly
when your test cases depend on the standard library. We've previously
worked around this by running tests in parallel, but that resulted in
several gigabytes of memory usage.

In contrast, our runner has been designed to run many analyzers on
many packages at once, efficiently traversing the dependency graph,
and caching data on disk.

To use the runner for our tests, we've added a "test mode", which
causes the runner to collect additional information that is required
to run tests. This includes extracting "// want" comments and making
a serialized form of facts accessible.

For the 'staticcheck' package, on a 16-core AMD 3950x system, wall
time and peak RSS have dropped from 12s and 7.9 GB to 7.5s and 200 MB.
In comparison, using x/tools/go/analysis without parallelism resulted
in 55s and 1.3 GB.

Closes gh-758