lintcmd/runner: don't create temporary files
I have no idea why we used temporary files in the first place, but
there's no point in doing it. Most people use tmpfs, so this doesn't
lower memory usage. It's also not a lot of data to begin with. When
checking the standard library, a total of 8.5 MB get allocated. It's
not even worth using a sync.Pool for it.