analysis/lint/testutil: don't reuse Staticcheck cache
go test has its own cache, which is keyed by all the inputs to a test,
including files that the process reads. However, if the runner uses the
Staticcheck cache, it may not end up reading files in testdata[1],
causing 'go test' to compute the wrong cache key. Subsequent changes to
files in testdata won't be picked up by 'go test', and it will report
stale results.

[1]: The files still end up being read, of course, but not by the test
process itself, but by processes it shells out to, such as 'go list'.
These reads are invisible to 'go test', and won't make it into the cache key.