Don't litter plain-text documentation with backticks
Check titles and descriptions contain a lot of backticks for Markdown
formatting, and treating them as plain text - for example for terminal
output - leads to a poor reading experience.

This change stores two versions of titles and descriptions. One
optimized for plain-text consumption, with backticks removed, and one
for Markdown consumption, such as the website.

We also introduce a new escape sequence, \", which renders as a
backtick in the Markdown version, but as a single quote in the plain
text version.