Implement binary output and merge mode
We implement a merge mode, which takes the results of multiple runs
and merges them. On an analysis-by-analysis basis we decide whether a
diagnostic has to occur in all runs to be merged. In addition, we take
into consideration whether a file was part of a run in the first
place, to avoid incorrectly dropping diagnostics.

We add a new output format called "binary", which writes diagnostics
and metadata using gob. This data is the input to the merge mode.

Currently, the metadata includes a list of all files that were
checked. In the future, we might change to amore finegrained data,
such as which lines of code were reachable in the IR.

Updates gh-938