lintcmd: add -matrix flag
When using the -matrix flag, Staticcheck will read a list of build
configurations from stdin, execute all of them, and merge their
results similar to -merge. This simplifies checking multiple
configurations on the same host.

When using build configurations, the plain output will annotate
diagnostics with the names of build configurations that produced them.

The output of a -matrix run can be used together with -f binary and
passed to -merge. This allows running multiple configurations per host
and merging them.

For now, build configurations can specify any environment variables
and flags, without restrictions. That's why they have to be provided
explicitly to Staticcheck and won't be read from staticcheck.conf. In
the future we may allow restricted build configurations to be read
automatically.

Updates gh-938